MappingBuilder, ESMappingBuilder : float par défaut pour decimal()
(c'est la valeur par défaut dans elasticsearch)

+++ class/MappingBuilder.php
@@ -2,7 +2,7 @@
 /**
  * This file is part of the "Docalist Search" plugin.
  *
- * Copyright (C) 2012-2015 Daniel Ménard
+ * Copyright (C) 2012-2016 Daniel Ménard
  *
  * For copyright and license information, please view the
  * LICENSE.txt file that was distributed with this source code.
@@ -123,7 +123,7 @@ interface MappingBuilder
      *
      * @return self
      */
-    public function decimal($type = 'double');
+    public function decimal($type = 'float');
 
     /**
      * Indique que le champ est de type date.
